+++ src/hooks/useBlockList.ts
@@ -1,10 +1,11 @@
 import ConnectyCube from "connectycube";
-import { useEffect, useState } from "react";
+import { useEffect, useState, useRef } from "react";
 
-export const BLOCK_LIST_NAME = "@ConnectyCube/blockList";
+export const BLOCK_LIST_LOG_TAG = "[useChat][useBlockList]";
+export const BLOCK_LIST_NAME = "ConnectyCubeBlockList";
 
 export type BlockListHook = {
-  blockedUsers: Set<number>;
+  blockedUsers: number[];
   isBlockedUser: (userId: number) => boolean;
   unblockUser: (userId: number) => Promise<void>;
   blockUser: (userId: number) => Promise<void>;
@@ -16,32 +17,34 @@ enum BlockAction {
 }
 
 function useBlockList(isConnected: boolean): BlockListHook {
+  let isApplied = useRef<boolean>(false).current;
+
   const [state, setState] = useState<Set<number>>(new Set<number>());
 
-  const addToState = (userId: number): void => {
-    setState((state) => {
-      state.add(userId);
-      return state;
-    });
-  };
+  const updateState = (userId: number, action: BlockAction) => {
+    const newState = new Set(state);
 
-  const deleteFromState = (userId: number): void => {
-    setState((state) => {
-      state.delete(userId);
-      return state;
-    });
+    if (action === BlockAction.DENY) {
+      newState.add(userId);
+    } else if (action === BlockAction.ALLOW) {
+      newState.delete(userId);
+    }
+
+    setState(newState);
   };
 
   const isBlocked = (userId: number): boolean => state.has(userId);
 
   const fetch = async (): Promise<void> => {
     if (!isConnected) {
+      console.warn(`${BLOCK_LIST_LOG_TAG}[fetch]: chat is not connected`);
       return;
     }
 
-    const blockList = await ConnectyCube.chat.privacylist.getList(BLOCK_LIST_NAME);
+    const blackListNames = await ConnectyCube.chat.privacylist.getNames();
 
-    if (blockList?.name === BLOCK_LIST_NAME) {
+    if (blackListNames.default === BLOCK_LIST_NAME) {
+      const blockList = await ConnectyCube.chat.privacylist.getList(BLOCK_LIST_NAME);
       const newState = blockList.items.reduce((list: Set<number>, item) => {
         if (item.action === BlockAction.DENY) {
           list.add(+item.user_id);
@@ -49,55 +52,60 @@ function useBlockList(isConnected: boolean): BlockListHook {
         return list;
       }, new Set<number>());
 
+      isApplied = true;
+
       setState(newState);
+    }
+  };
+
+  const upsert = async (user_id: number, action: BlockAction): Promise<void> => {
+    if (!isConnected) {
+      console.warn(`${BLOCK_LIST_LOG_TAG}[upsert]: ${action} user ${user_id} failed, chat is not connected`);
+      return;
+    }
+
+    const blockList = {
+      name: BLOCK_LIST_NAME,
+      items: [{ user_id, action, mutualBlock: true }],
+    };
+
+    if (isApplied) {
+      await ConnectyCube.chat.privacylist.update(blockList);
     } else {
-      await ConnectyCube.chat.privacylist.create({
-        name: BLOCK_LIST_NAME,
-        items: [],
-      });
+      await ConnectyCube.chat.privacylist.create(blockList);
       await ConnectyCube.chat.privacylist.setAsDefault(BLOCK_LIST_NAME);
     }
   };
 
-  const update = async (user_id: number, action: BlockAction): Promise<void> => {
-    if (!isConnected) {
-      return;
-    }
-
-    if (action === BlockAction.ALLOW && !isBlocked(user_id)) {
-      console.warn("[useChat][useBlockList][update]: user is not blocked");
-      return;
-    }
-
-    if (action === BlockAction.DENY && isBlocked(user_id)) {
-      console.warn("[useChat][useBlockList][update]: user is already blocked");
-      return;
-    }
-
-    return ConnectyCube.chat.privacylist.update({
-      name: BLOCK_LIST_NAME,
-      items: [{ user_id, action, mutualBlock: true }],
-    });
-  };
-
   const unblock = async (userId: number): Promise<void> => {
-    await update(userId, BlockAction.ALLOW);
-    deleteFromState(userId);
+    if (!isBlocked(userId)) {
+      console.warn(`${BLOCK_LIST_LOG_TAG}[unblock]: user ${userId} is not blocked`);
+      return;
+    }
+
+    await upsert(userId, BlockAction.ALLOW);
+    updateState(userId, BlockAction.ALLOW);
   };
 
   const block = async (userId: number): Promise<void> => {
-    await update(userId, BlockAction.DENY);
-    addToState(userId);
+    if (isBlocked(userId)) {
+      console.warn(`${BLOCK_LIST_LOG_TAG}[block]: user ${userId} is already blocked`);
+      return;
+    }
+
+    await upsert(userId, BlockAction.DENY);
+    updateState(userId, BlockAction.DENY);
   };
 
   useEffect(() => {
+    console.warn("useEffect", { isConnected });
     if (isConnected) {
       fetch();
     }
   }, [isConnected]);
 
   return {
-    blockedUsers: state,
+    blockedUsers: Array.from(state),
     isBlockedUser: isBlocked,
     unblockUser: unblock,
     blockUser: block,